[DOC] Recommendation when using a lot of metrics filters
Following a report from a user, we discovered it might be necessary to increase `Xss` in `jvm.options`.

The (third party) `Metriks::ExponentiallyDecayingSample` uses a red-black tree to ensure the worst-case cost to look up an individual entry is `O(log N)`. It keeps a maximum of 1024 entries around, which means that the worst-case lookup is 10 steps.